requirements.txt
ton-client-py
fastapi
uvicorn[standard]
python-jose[cryptography]
aiofiles
httpx
